separate file for launch request handler
