+++ lambda/custom/utils.js
@@ -1,0 +1,9 @@
+const supportsAPL = (handlerInput) => {
+	const supportedInterfaces = handlerInput.requestEnvelope.context.System.device.supportedInterfaces;
+	const aplInterface = supportedInterfaces['Alexa.Presentation.APL'];
+	return aplInterface != null && aplInterface != undefined;
+}
+
+module.exports = {
+    supportsAPL
+}
